Fix unsigned vs signed comparison warnings in basic.c.
Make variableSet* update the variable list rather than returning the
new list. This avoids forgetting to assign the result, which I've done
a couple of times already.

In set-variable, don't pre-insert the old value of the variable into
the minibuffer. Emacs doesn't, and it's irritating.

+++ src/extern.h
@@ -4,7 +4,7 @@ size_t get_goalc_wp(Window *wp);
 size_t get_goalc(void);
 int backward_char(void);
 int forward_char(void);
-void goto_line(int to_line);
+void goto_line(size_t to_line);
 void gotobob(void);
 void gotoeob(void);
 int next_line(void);
